+++ src/bssunfold/core/__init__.py
@@ -24,6 +24,9 @@ from .unfold_reconst import solve_reconst
 from .unfold_scipy_direct_method import solve_scipy_direct
 from .unfold_tsvd import solve_tsvd
 from .unfold_lanczos import solve_lanczos
+from .unfold_cgls import solve_cgls
+from .unfold_gks import solve_gks
+from .unfold_tikhonov_tv import solve_tikhonov_tv
 from .unfold_parametric2 import solve_parametric2
 from .unfold_fruit_like import solve_fruit_like
 from .unfold_hybrid_parametric import solve_hybrid_parametric
@@ -70,6 +73,9 @@ from .unfold_reconst import unfold_reconst
 from .unfold_scipy_direct_method import unfold_scipy_direct_method
 from .unfold_tsvd import unfold_tsvd
 from .unfold_lanczos import unfold_lanczos
+from .unfold_cgls import unfold_cgls
+from .unfold_gks import unfold_gks
+from .unfold_tikhonov_tv import unfold_tikhonov_tv
 from .unfold_parametric2 import unfold_parametric2
 from .unfold_smt import unfold_smt
 from .unfold_scip import unfold_scip
@@ -107,6 +113,9 @@ __all__ = [
     "solve_scipy_direct",
     "solve_tsvd",
     "solve_lanczos",
+    "solve_cgls",
+    "solve_gks",
+    "solve_tikhonov_tv",
     "solve_parametric2",
     "solve_fruit_like",
     "solve_hybrid_parametric",
@@ -146,6 +155,9 @@ __all__ = [
     "unfold_scipy_direct_method",
     "unfold_tsvd",
     "unfold_lanczos",
+    "unfold_cgls",
+    "unfold_gks",
+    "unfold_tikhonov_tv",
     "unfold_parametric2",
     "unfold_fruit_like",
     "unfold_hybrid_parametric",